refactor(orchestrator): extract tier-migration cycle into TierMigrationCoordinator (#1526)
Behavior-preserving extraction of the hot/cold tier-migration lifecycle
(config gating, in-flight guard, min-interval throttle, budget selection,
candidate scanning, promotion/demotion execution, status recording) from
orchestrator.ts into orchestration/tier-migration-coordinator.ts.

The orchestrator constructs one TierMigrationCoordinator and keeps thin
delegating methods (runTierMigrationCycle / runTierMigrationNow /
getTierMigrationStatus) so call sites and the public API are unchanged.
The in-flight guard and last-run timestamp move to the coordinator;
cold-storage construction is injected (createColdStorage) so the module
owns no direct storage.ts import (#1533 ratchet: directStorageImports
unchanged at 38).

orchestrator.ts: 20563 -> 20386 LOC (ratchet baseline tightened).
entity-hardening 246 green; preflight:quick green.

Chokepoints used: none new (extract-when-touched move; no feature).
